change required to work with PSoC creator

+++ sw_modules/median_filter/median_filter.h
@@ -24,7 +24,7 @@
  * different types simultaneously. This is limitation of C language, that lacking templates, ...
  * or my imaginations :)
  */
-typedef uint16_t median_value_t; 
+typedef uint32_t median_value_t; 
 
 //=================================================================================================
 // data types
